CON-630: Move dag-related code under `casper.dag`.
