+++ casper/src/test/scala/io/casperlabs/casper/dag/CliqueTest.scala
@@ -1,4 +1,4 @@
-package io.casperlabs.casper.util
+package io.casperlabs.casper.dag
 
 import io.casperlabs.casper.helper.BlockGenerator
 import org.scalatest.{Assertion, FlatSpec, Matchers}
